Add download service to pki service deps
